requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
decorator==4.4.2
flatbuffers==24.3.25
frozenlist==1.4.1
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
magic-filter==1.0.12
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pdf417decoder==1.0.8
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
protobuf==5.27.3
pydantic==2.8.2
pydantic_core==2.20.1
PyMatting==1.1.12
pypng==0.20220715.0
pyreadline3==3.4.1
pyzbar==0.1.9
qrcode==7.4.2
qrcode-artistic==3.0.2
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rpds-py==0.19.1
scikit-image==0.24.0
scipy==1.14.0
segno==1.6.1
sympy==1.13.1
tifffile==2024.7.24
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
zxing-cpp==2.2.0
